+++ src/CMakeLists.txt
@@ -19,4 +19,5 @@ add_subdirectory(init)
 add_subdirectory(ipc)
 add_subdirectory(lib)
 add_subdirectory(math)
+add_subdirectory(schedule)
 add_subdirectory(tasks)
