+++ src/arch/xtensa/smp/init.c
@@ -21,7 +21,7 @@
 #include <sof/lock.h>
 #include <sof/notifier.h>
 #include <sof/panic.h>
-#include <sof/schedule.h>
+#include <sof/schedule/schedule.h>
 #include <sof/task.h>
 #include <platform/idc.h>
 #include <stdint.h>
