+++ src/arch/xtensa/smp/schedule.c
@@ -12,7 +12,7 @@
 
 #include <xtos-structs.h>
 #include <arch/cpu.h>
-#include <sof/schedule.h>
+#include <sof/schedule/schedule.h>
 
 struct schedule_data **arch_schedule_get_data(void)
 {
