+++ src/include/sof/ssp.h
@@ -11,7 +11,7 @@
 #include <sof/dai.h>
 #include <sof/io.h>
 #include <sof/lock.h>
-#include <sof/schedule.h>
+#include <sof/schedule/schedule.h>
 #include <sof/trace.h>
 #include <sof/wait.h>
 #include <platform/clk-map.h>
