+++ src/schedule/CMakeLists.txt
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+add_local_sources(sof edf_schedule.c ll_schedule.c schedule.c)
