+++ test/cmocka/src/audio/pipeline/pipeline_connect_upstream.c
@@ -7,7 +7,7 @@
 #include <stdint.h>
 #include <sof/audio/component.h>
 #include <sof/audio/pipeline.h>
-#include <sof/edf_schedule.h>
+#include <sof/schedule/edf_schedule.h>
 #include "pipeline_mocks.h"
 #include "pipeline_connection_mocks.h"
 #include <stdarg.h>
