+++ test/cmocka/src/audio/pipeline/pipeline_new.c
@@ -7,7 +7,7 @@
 #include <stdint.h>
 #include <sof/audio/component.h>
 #include <sof/audio/pipeline.h>
-#include <sof/schedule.h>
+#include <sof/schedule/schedule.h>
 #include "pipeline_mocks.h"
 #include <stdarg.h>
 #include <stddef.h>
