+++ test/cmocka/src/audio/pipeline/pipeline_new_allocation_mocks.c
@@ -6,7 +6,7 @@
 
 #include <sof/audio/component.h>
 #include <sof/audio/pipeline.h>
-#include <sof/edf_schedule.h>
+#include <sof/schedule/edf_schedule.h>
 #include <stdarg.h>
 #include <stddef.h>
 #include <setjmp.h>
